+++ src/main/java/se/kth/edge/Edge.java
@@ -32,7 +32,7 @@ public class Edge {
         if (edgeId != eId) {
             // TODO this is an update from other edges
         }
-        int estimatedRate = wMonitor.getExpectedKeyArrivalRate(kid);
+        float estimatedRate = wMonitor.getExpectedKeyArrivalRate(kid);
         if (coordinators.containsKey(kid) && coordinators.get(kid) == eId) {
             cache.insertPriorityKey(kid, time, estimatedRate); // prioritize keys that the edge is its coordinator.
         } else {
